+++ src/components/profile/ProfileName/index.tsx
@@ -1,22 +1,11 @@
 import React, { useState } from "react";
 import { View, Text, TouchableOpacity } from "react-native";
+import PropTypes from "prop-types";
 import { AntDesign } from "@expo/vector-icons";
-import { QrCodeGenerator } from "../../QRGenerator/QrGenerator";
+import { QrCodeGenerator } from "../../QrGenerator";
 import { styles } from "./ProfileNameStyles";
 
-interface IProfileNameProps {
-  fin: string;
-  name: string;
-  navigation: any;
-  isPreview: boolean;
-}
-
-export const ProfileName = ({
-  fin,
-  name,
-  navigation,
-  isPreview
-}: IProfileNameProps) => {
+export const ProfileName = ({ status, photo, fin, name, isPreview }) => {
   const [isDialogVisible, setVisible] = useState(false);
 
   const toggleVisibility = () => {
@@ -27,18 +16,7 @@ export const ProfileName = ({
     <View style={styles.textContainer}>
       <Text style={styles.fin}>{fin}</Text>
       <Text style={styles.name}>{name}</Text>
-      {isPreview ? ( // Is Preview Section
-        <View style={styles.previewContainer}>
-          <Text style={styles.previewText}>Preview Only</Text>
-          <TouchableOpacity
-            style={styles.previewButton}
-            onPress={navigation.goBack}
-          >
-            <AntDesign name="back" size={15} />
-            <Text style={styles.previewButtonText}>Go back to QR scanner</Text>
-          </TouchableOpacity>
-        </View>
-      ) : (
+      {!isPreview && status === 1 ? ( // Is Preview Section
         // Not Preview Section
         <TouchableOpacity
           style={styles.shareContainer}
@@ -47,11 +25,22 @@ export const ProfileName = ({
           <AntDesign name="qrcode" size={15} color="#808080" />
           <Text style={styles.shareText}>SHARE ID</Text>
           <QrCodeGenerator
+            photo={photo}
+            name={name}
+            fin={fin}
             isVisible={isDialogVisible}
             handleCancel={toggleVisibility}
           />
         </TouchableOpacity>
-      )}
+      ) : null}
     </View>
   );
+};
+
+ProfileName.propTypes = {
+  fin: PropTypes.string,
+  name: PropTypes.string,
+  photo: PropTypes.string,
+  isPreview: PropTypes.bool,
+  status: PropTypes.number
 };
